+++ hw12_13_14_15_calendar/internal/config/calendar/config.go
@@ -1,4 +1,4 @@
-package main
+package calendar
 
 import (
 	"fmt"
@@ -6,9 +6,6 @@ import (
 	"github.com/BurntSushi/toml"
 )
 
-// При желании конфигурацию можно вынести в internal/config.
-// Организация конфига в main принуждает нас сужать API компонентов, использовать
-// при их конструировании только необходимые параметры, а также уменьшает вероятность циклической зависимости.
 type Config struct {
 	Logger  LoggerConf
 	Storage StorageConf
@@ -38,7 +35,7 @@ type GRPCConf struct {
 	Port string
 }
 
-func NewConfig() Config {
+func New() Config {
 	return Config{}
 }
