+++ hw12_13_14_15_calendar/internal/consumer/message.go
@@ -1,0 +1,5 @@
+package consumer
+
+type Message struct {
+	Data []byte
+}
